dts: refine pre-test setup and teardown steps
The major part is the removal of _set_up_execution() and
_tear_down_execution() of Node in lieu of using super() in the
superclasses, which simplifies the code while achieving the same thing.

The minor changes are the move of virtual devices and build target
setup/teardown into SutNode from Node since both are DPDK-related which
are only going to run on the SutNode.

Signed-off-by: Juraj Linkeš <juraj.linkes@pantheon.tech>
Reviewed-by: Luca Vizzarro <luca.vizzarro@arm.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>