net/tap: simplify internals
The names of Linux network devices are IFNAMSIZ(16) not the
same as DPDK which has up to 64 characters. Don't need to
hold onto the whole ifreq to save the remote interface flags.

Make sure packet and byte counters are read once, so that global
and per-queue values add up. No need for separate rx_nombuf counter
since there is an alloc_failed value in ethdev.

Signed-off-by: Stephen Hemminger <stephen@networkplumber.org>
Signed-off-by: 0-day Robot <robot@bytheb.org>