net/ixgbe/base: fix crash while loading driver
Issue observed: system crash While loading the driver when debug macros
were defined.

The solution removes the debug macro call from
ixgbe_set_mac_type() function which causes the crash.

So far, the ixgbe_set_mac_type() function has called in the ixgbe driver
ixgbe_probe() function before the struct ixgbe_hw is properly defined.
In this case, the field back of the structure that must contain
pointer to an adapter has not yet been defined and should not be used
to obtain driver information which is needed in debug macros.
Moreover bearing in mind the recommendations to not print out excessive
amount of information it was decided to remove the debug information.

Signed-off-by: Slawomir Mrozowicz <slawomirx.mrozowicz@intel.com>
Reviewed-by: Tyl, RadoslawX <radoslawx.tyl@intel.com>
Reviewed-by: Skajewski, PiotrX <piotrx.skajewski@intel.com>
Tested-by: nosbuild <nosbuild@intel.com>
Reviewed-by: Michael, Alice <alice.michael@intel.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>