graph: fix does not return the unique id when create graph
When the order of graph destroy is not the reverse order of create,
that is, when it is destroyed at will, the newly created graph id will
be the same as the existing graph id, which is not the expected unique
graph id. This graph id incorrectly corresponds to multiple graphs.

Fixes: a91fecc19c5c ("graph: implement create and destroy")
Cc: stable@dpdk.org

Reported-by: Fan Yin <yinf1@chinatelecom.cn>
Signed-off-by: Gongming Chen <chengm11@chinatelecom.cn>
Signed-off-by: Fan Yin <yinf1@chinatelecom.cn>
Signed-off-by: 0-day Robot <robot@bytheb.org>